+++ iddd_collaboration/src/main/java/com/saasovation/collaboration/port/adapter/persistence/view/MySQLCalendarProjection.java
@@ -30,17 +30,10 @@ import com.saasovation.common.event.sourcing.EventDispatcher;
 import com.saasovation.common.port.adapter.persistence.AbstractProjection;
 import com.saasovation.common.port.adapter.persistence.ConnectionProvider;
 
-public class MySQLCalendarProjection
-        extends AbstractProjection
-        implements EventDispatcher {
+public class MySQLCalendarProjection extends AbstractProjection implements EventDispatcher {
 
-    private static final Class<?> understoodEventTypes[] = {
-        CalendarCreated.class,
-        CalendarDescriptionChanged.class,
-        CalendarRenamed.class,
-        CalendarShared.class,
-        CalendarUnshared.class
-    };
+    private static final Class<?> understoodEventTypes[] = {CalendarCreated.class, CalendarDescriptionChanged.class, CalendarRenamed.class, CalendarShared.class,
+                                                            CalendarUnshared.class};
 
     public MySQLCalendarProjection(EventDispatcher aParentEventDispatcher) {
         super();
@@ -60,30 +53,19 @@ public class MySQLCalendarProjection
 
     @Override
     public boolean understands(DispatchableDomainEvent aDispatchableDomainEvent) {
-        return this.understandsAnyOf(
-                aDispatchableDomainEvent.domainEvent().getClass(),
-                understoodEventTypes);
+        return this.understandsAnyOf(aDispatchableDomainEvent.domainEvent().getClass(), understoodEventTypes);
     }
 
     protected void when(CalendarCreated anEvent) throws Exception {
         Connection connection = ConnectionProvider.connection();
 
         // idempotent operation
-        if (this.exists(
-                "select calendar_id from tbl_vw_calendar "
-                    + "where tenant_id = ? and calendar_id = ?",
-                anEvent.tenant().id(),
-                anEvent.calendarId().id())) {
+        if (this.exists("select calendar_id from tbl_vw_calendar " + "where tenant_id = ? and calendar_id = ?", anEvent.tenant().id(), anEvent.calendarId().id())) {
             return;
         }
 
-        PreparedStatement statement =
-                connection.prepareStatement(
-                        "insert into tbl_vw_calendar("
-                        + "calendar_id, description, name, "
-                        + "owner_email_address, owner_identity, owner_name, "
-                        + "tenant_id"
-                        + ") values(?,?,?,?,?,?,?)");
+        PreparedStatement statement = connection.prepareStatement(
+                "insert into tbl_vw_calendar(" + "calendar_id, description, name, " + "owner_email_address, owner_identity, owner_name, " + "tenant_id" + ") values(?,?,?,?,?,?,?)");
 
         statement.setString(1, anEvent.calendarId().id());
         statement.setString(2, anEvent.description());
@@ -103,10 +85,7 @@ public class MySQLCalendarProjection
     protected void when(CalendarDescriptionChanged anEvent) throws Exception {
         Connection connection = ConnectionProvider.connection();
 
-        PreparedStatement statement =
-                connection.prepareStatement(
-                        "update tbl_vw_calendar set description=? "
-                        + "where calendar_id = ?");
+        PreparedStatement statement = connection.prepareStatement("update tbl_vw_calendar set description=? " + "where calendar_id = ?");
 
         statement.setString(1, anEvent.description());
         statement.setString(2, anEvent.calendarId().id());
@@ -117,10 +96,7 @@ public class MySQLCalendarProjection
     protected void when(CalendarRenamed anEvent) throws Exception {
         Connection connection = ConnectionProvider.connection();
 
-        PreparedStatement statement =
-                connection.prepareStatement(
-                        "update tbl_vw_calendar set name=? "
-                        + "where calendar_id = ?");
+        PreparedStatement statement = connection.prepareStatement("update tbl_vw_calendar set name=? " + "where calendar_id = ?");
 
         statement.setString(1, anEvent.name());
         statement.setString(2, anEvent.calendarId().id());
@@ -135,10 +111,7 @@ public class MySQLCalendarProjection
     protected void when(CalendarUnshared anEvent) throws Exception {
         Connection connection = ConnectionProvider.connection();
 
-        PreparedStatement statement =
-                connection.prepareStatement(
-                        "delete from tbl_vw_calendar_sharer "
-                        + "where tenant_id=? and calendar_id=? and participant_identity=?");
+        PreparedStatement statement = connection.prepareStatement("delete from tbl_vw_calendar_sharer " + "where tenant_id=? and calendar_id=? and participant_identity=?");
 
         statement.setString(1, anEvent.tenant().id());
         statement.setString(2, anEvent.calendarId().id());
@@ -147,31 +120,18 @@ public class MySQLCalendarProjection
         this.execute(statement);
     }
 
-    private void insertCalendarSharer(
-            Tenant aTenant,
-            CalendarId aCalendarId,
-            CalendarSharer aCalendarSharer)
-    throws Exception {
+    private void insertCalendarSharer(Tenant aTenant, CalendarId aCalendarId, CalendarSharer aCalendarSharer) throws Exception {
 
         Connection connection = ConnectionProvider.connection();
 
         // idempotent operation
-        if (this.exists(
-                "select id from tbl_vw_calendar_sharer "
-                    + "where tenant_id = ? and calendar_id = ? and participant_identity = ?",
-                aTenant.id(),
-                aCalendarId.id(),
-                aCalendarSharer.participant().identity())) {
+        if (this.exists("select id from tbl_vw_calendar_sharer " + "where tenant_id = ? and calendar_id = ? and participant_identity = ?", aTenant.id(), aCalendarId.id(),
+                        aCalendarSharer.participant().identity())) {
             return;
         }
 
-        PreparedStatement statement =
-                connection.prepareStatement(
-                        "insert into tbl_vw_calendar_sharer("
-                        + "id, calendar_id, "
-                        + "participant_email_address, participant_identity, participant_name, "
-                        + "tenant_id"
-                        + ") values(?,?,?,?,?,?)");
+        PreparedStatement statement = connection.prepareStatement(
+                "insert into tbl_vw_calendar_sharer(" + "id, calendar_id, " + "participant_email_address, participant_identity, participant_name, " + "tenant_id" + ") values(?,?,?,?,?,?)");
 
         statement.setLong(1, 0);
         statement.setString(2, aCalendarId.id());
